+++ packages/providers/lib.esm/fallback-provider.js
@@ -15,6 +15,7 @@ import { deepCopy, defineReadOnly, shallowCopy } from "@ethersproject/properties
 import { shuffled } from "@ethersproject/random";
 import { poll } from "@ethersproject/web";
 import { BaseProvider } from "./base-provider";
+import { isCommunityResource } from "./formatter";
 import { Logger } from "@ethersproject/logger";
 import { version } from "./_version";
 const logger = new Logger(version);
@@ -354,14 +355,16 @@ export class FallbackProvider extends BaseProvider {
         }
         const providerConfigs = providers.map((configOrProvider, index) => {
             if (Provider.isProvider(configOrProvider)) {
-                return Object.freeze({ provider: configOrProvider, weight: 1, stallTimeout: 750, priority: 1 });
+                const stallTimeout = isCommunityResource(configOrProvider) ? 2000 : 750;
+                const priority = 1;
+                return Object.freeze({ provider: configOrProvider, weight: 1, stallTimeout, priority });
             }
             const config = shallowCopy(configOrProvider);
             if (config.priority == null) {
                 config.priority = 1;
             }
             if (config.stallTimeout == null) {
-                config.stallTimeout = 750;
+                config.stallTimeout = isCommunityResource(configOrProvider) ? 2000 : 750;
             }
             if (config.weight == null) {
                 config.weight = 1;