+++ packages/web/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "web/5.0.8";
+export declare const version = "web/5.0.9";